backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
pandas==2.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1 